Unify tests using Ray, fix tabulate
