+++ tests/full_wlm/with_ray/test_ray_batch.py
@@ -2,12 +2,12 @@ import logging
 import sys
 import time
 from os import environ
-from shutil import which
 
 import pytest
 
 from smartsim import Experiment
 from smartsim.exp.ray import RayCluster
+from smartsim.launcher import slurm
 
 """Test Ray cluster Slurm launch and shutdown.
 """
@@ -15,7 +15,6 @@ from smartsim.exp.ray import RayCluster
 # retrieved from pytest fixtures
 if pytest.test_launcher not in pytest.wlm_options:
     pytestmark = pytest.mark.skip(reason="Not testing WLM integrations")
-
 
 environ["OMP_NUM_THREADS"] = "1"
 shouldrun = True
@@ -33,11 +32,11 @@ pytestmark = pytest.mark.skipif(
 
 def test_ray_launch_and_shutdown_batch(fileutils, wlmutils, caplog):
     launcher = wlmutils.get_test_launcher()
-    if launcher != "pbs":
-        pytest.skip("Test only runs on systems with PBS as WLM")
+    if launcher == "local":
+        pytest.skip("Test cannot be run with local launcher")
 
     caplog.set_level(logging.CRITICAL)
-    test_dir = fileutils.make_test_dir("test-ray-pbs-launch-and-shutdown-batch")
+    test_dir = fileutils.make_test_dir("test-ray-launch-and-shutdown-batch")
 
     exp = Experiment("ray-cluster", test_dir, launcher=launcher)
     cluster = RayCluster(
@@ -46,17 +45,16 @@ def test_ray_launch_and_shutdown_batch(fileutils, wlmutils, caplog):
         ray_args={"num-cpus": 4},
         launcher=launcher,
         num_nodes=2,
-        alloc=None,
         batch=True,
-        ray_port=6830,
-        time="00:05:00",
-        interface=wlmutils.get_test_interface()
-        )
+        interface=wlmutils.get_test_interface(),
+        batch_args={"A": wlmutils.get_test_account(),
+                    "queue": "debug-flat-quad"} if launcher=='cobalt' else None,
+    )
 
     exp.generate(cluster)
     exp.start(cluster, block=False, summary=True)
-    ctx = ray.init("ray://" + cluster.get_head_address() + ":10001")    
-    
+    ctx = ray.client("ray://" + cluster.get_head_address() + ":10001").connect()
+
     right_resources = False
     trials = 10
     while not right_resources and trials > 0:
